Timeline clean ups
- Unify the timeline time unit to be long (milliseconds), to save the
  back and forth conversion between seconds/mills. Note that some parts
  of the code still use float, but that is out of the scope.
- Remove a useless interface (ITimelineKeyframesInterface).
- Remove unused code plus some tweaks here and there.

Task-number: QT3DS-2850

+++ src/Authoring/Studio/Palettes/TimelineGraphicsView/TimelineControl.h
@@ -53,8 +53,8 @@ private:
     TimelineGraphicsScene *m_scene = nullptr;
     RowTimeline *m_rowTimeline = nullptr;
     ITimelineTimebar *m_timebar = nullptr;
-    double m_startTime = 0;
-    double m_endTime = 0;
+    long m_startTime = 0;
+    long m_endTime = 0;
 };
 
 #endif // TIMELINECONTROL_H
